Swap command - Remove unneeded lines - Add modcommands
